CL-399 Refactor to new UserSlugService
